requirements.txt
arrow==0.4.2
awesome-slugify==1.5
bottle==0.12.9
fake-factory==0.4.0
Jinja2==2.7.3
MarkupSafe==0.23
psutil==3.2.2
python-dateutil==2.2
redis==2.10.5
redislite==1.0.257
regex==2015.11.14
six==1.7.3
Unidecode==0.4.16
wheel==0.24.0
